requirements.txt
django==4.1.0
uwsgi==2.0.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
#mysqlclient==1.4.6